Create Beautiful Binary String
Alice has a binary string,B , of length n. She thinks a binary string is beautiful if and only if it doesn't contain the substring "010".

In one step, Alice can change a0 to a1 (or vice-versa). Count and print the minimum number of steps needed to make Alice see the string as beautiful.